// src/app/interfaces/laboratory.interface.ts
export interface Laboratory {
  actividad: string;
  ano_de_fundacion?: string;
  area_conocimiento?: string;
  capacidad_estudiantes?: number;
  capacidad_maxima_estudiantes: number;
  cargo?: string;
  categoria_docente?: string;
  celular?: string;
  cod_escuela: number;
  cod_georef: number;
  cod_lab: number;
  direccion: string;
  email_universidad?: string;
  escuela_o_departamento: string;
  espacio: string;
  hora_semana?: string;
  identificacion?: string;
  nombre_de_laboratorio: string;
  nombre_responsable?: string;
  resolucion_de_creacion: string;
  telefono_lab?: string;
  telefono_responsable?: string;
  key$?:string
}